# Conflicts:
#	src/Core/Presentation/ViewModels/Rating/Items/RatingItemViewModel.cs
#	src/Core/Presentation/ViewModels/Rating/RatingViewModel.cs
#	src/Droid/PrankChat.Mobile.Droid.csproj
#	src/iOS/AppTheme/Theme.cs
#	src/iOS/PrankChat.Mobile.iOS.csproj

+++ src/Core/Models/Api/OrderApiModel.cs
@@ -12,7 +12,7 @@ namespace PrankChat.Mobile.Core.Models.Api
         public int Id { get; set; }
 
         [JsonProperty("price")]
-        public long? Price { get; set; }
+        public double? Price { get; set; }
 
         [JsonProperty("title")]
         public string Title { get; set; }
